Move all units (500km at the east)

+++ GameShared/Game/GPSPosition.cs
@@ -4,22 +4,16 @@
 
 namespace GameShared.Game
 {
-    using System;
-    using System.Collections.Generic;
-    using System.Linq;
-    using System.Text;
-    using System.Threading.Tasks;
-
     public struct GPSPosition : IPosition
     {
-        public GPSPosition(float latitude, float longitude)
+        public GPSPosition(double latitude, double longitude)
         {
             Latitude = latitude;
             Longitude = longitude;
         }
 
-        public float Latitude { get; set; }
+        public double Latitude { get; set; }
 
-        public float Longitude { get; set; }
+        public double Longitude { get; set; }
     }
 }
